audacity: fix unclean shutdown due to sqlite error
This applies a patch from upstream[1] that replaces the previously used
method of using the SQLITE_DBPAGE virtual table extension.

This fixes audacity hanging when closing it, which leads to it having to
be killed. That results in incompletely saved project files which
triggers the file recovery dialogue on every start.

Fixes #130347.

[1] https://github.com/audacity/audacity/pull/1802

(cherry picked from commit fed4002ba3930bf16573fafec42226ac52bd926c)