matplotlib-inline: Don't depend on matplotlib -> smaller ipython closure
The lib `matplotlib-inline` is only a shim to integrate `matplotlib` in
ipython, but it doesn't actually require it.

If one wants to use the matplotlib integration, `matplotlib` need to be
installed separatedly, then calling `%matplotlib inline` at a ipython
prompt will work.

Not including `matplotlib` is closer to a non-Nix installation:
Installing ipython in a fresh venv also does not install `matplotlib`.

NOTE: ipython's closure is now ~160MB instead of ~400MB.
(including ~100MB for python itself)