tor-browser-bundle-bin: Keep files when store path has not changed
Originally, some TBB files in the user's $HOME are always deleted on
startup, and as a result, addons are not loaded automatically at
startup. For example, NoScript is disabled by default and needs a manual
reload cycle (dis-/enable by hand).

This change preserves the files with the addon information as long as
the contained store path is the same, i.e. TBB was not modified.  When
TBB is updated, the files are removed as before, so that they don't
contain paths to the store location of the (now outdated) previous TBB.
A manual reload addon cycle is necessary once.

(cherry picked from commit 8985aca004193bd179ab49e49d2eff37d77a44a5)