blightmud: init at 3.5.0
Blightmud is a terminal client for connecting to Multi User Dungeon
(MUD) games. It is written in Rust and supports TLS, GMCP, MSDP, MCCP2,
tab completion, text searching and a split view for scrolling. Blightmud
can be customized with Lua scripting for aliases, triggers, timers,
customized status bars, and more. Blightmud supports several
accessibility features including an optional built-in text-to-speech
engine and a screen reader friendly mode.

For nixpkgs it is largely a standard derivation for a rust project using
`rustPlatform.buildRustPackage`. There is some customization required
for the optional text-to-speech (TTS) engine support. In this case the
derivation must also set the `LIBCLANG_PATH` and customize
`BINDGEN_EXTRA_CLANG_ARGS` in order for a required crate to be able to
`rust-bindgen` the `libspeechd` dependency it wraps. Lastly the
derivation has to skip some integration-style tests that don't play
nicely with the nixpkgs build environment - the majority of unit tests
work so they are left running in the check phase.

Since the TTS support brings in heavy dependencies, but is a useful
accessibility feature, the Blightmud derivation is added to
`all-packages.nix` twice:

1. the `blightmud` attribute builds a configuration without TTS support.
2. the `blightmud-tts` attribute builds a configuration _with_ TTS
   support.

The new Blightmud derivation is placed in `pkgs/games/blightmud/`
following the precedent set by another packaged GUI-based MUD client,
`mudlet` with `pkgs/games/mudlet/`.

(cherry picked from commit ae1bee344a09129db2c13d5564e632934b68cdaf)