neomutt: apply patch for CVE-2022-1328
This fixes a buffer overflow in NeoMutt, to quote the original
report[1]:

> Hello, In mutt_decode_uuencoded(), the line length is read from the
> untrusted uuencoded part without validation. This could result in
> including private memory in message parts, for example fragments of
> other messages, passphrases or keys in replys.

Security advistory for the corresponding CVE-2022-1328 is on GitHub[2].

Applying the entire release 20220415 is IMHO too risky because too much
has changed, but applying the patch only works fine as well here.

In NeoMutt 20220415 / Mutt 2.2.3 there's also a fix for an integer
overflow[3] of a `strlen()` for very large messages which is however not
exploitable for NeoMutt according to the upstream maintainers[4], so I
won't backport this patch as well.

Related to https://github.com/NixOS/nixpkgs/pull/168800.

[1] https://gitlab.com/muttmua/mutt/-/issues/404
[2] https://github.com/advisories/GHSA-qfrq-pp74-gpff
[3] https://gitlab.com/muttmua/mutt/-/issues/405
[4] https://github.com/neomutt/neomutt/commit/2bedc9762e2be679a24c8af9c56d16d670aef958: