nextcloud-exporter: mark as vulnerable to CVE-2022-21698
See also #169928 for a PR that would fix this vulnerability; after a
brief discussion with @Ma27 we came to the conclusion that breaking
backwards compatibility silently is more of a pain than breaking
evaluation loudly. We consider the severity of this vulnerability as
relatively low, because prometheus exporters should usually not be
publicly accessible anyway, and believe that this is the best approach
to allowing people who run the exporter the choice of running the
compatible version.

Closes #169928.