grafana: fix CVE-2022-29170
We're still using 8.4.x on 21.11 because 8.5 contained a few breaking
changes[1]. The patch for 8.5[2] was almost compatible with 8.4, but had
to be modified a bit because `New` in `http_client_provider` had a
slightly different signature:

    func New(cfg *setting.Cfg, tracer tracing.Tracer, features featuremgmt.FeatureToggles) *sdkhttpclient.Provider

on 8.4.7 vs

    func New(cfg *setting.Cfg, tracer tracing.Tracer) *sdkhttpclient.Provider

on 8.5.2.

I only had to adjust the signature in the test (because `PluginRequestValidator`
was added), but nothing else because these calls are automatically
resolved via `wire`.

[1] https://github.com/grafana/grafana/releases/tag/v8.5.0
[2] https://github.com/grafana/grafana/commit/2f756845006820de4ad2b33e4be8338b81217d41