pkgsStatic.glib: fix build
Since 9f309c87925 ("stdenv: mesonFlags: use canExecute in needs_exe_wrapper"),
glib's tests are built (but not run by default) in pkgsStatic.  This
broke the build because some of the tests required shared libraries.
This patch (which is awaiting review upstream) disables building those
libraries and any tests that depend on them in a static-only build.