haskell.packages.ghc942: get inital support up
This change introduces some actual overrides in
configuration-ghc-9.4.x.nix (based on the ones for 9.2) with the aim of
getting a basic 9.4 package set up. In particular the following
“interesting” packages are working:

* hpack
* hoogle
* cabal2nix
* cabal-install

You can get a better overview by looking at the overrides and the
versionedCompilerJobs in release-haskell.nix. Next obvious steps would
be working towards basic HLS support, but it seems that basically none
of the plugins are working so far. Big blockers here:

* https://github.com/alanz/ghc-exactprint/issues/114
* https://github.com/ndmitchell/hlint/issues/1413
* https://github.com/haskell/haskell-language-server/issues/3190

We ourselves have some churn to take care of and should try to release
the changes in https://github.com/NixOS/cabal2nix/pull/571 soon.

Co-authored-by: Rebecca Turner <rbt@sent.as>
Co-authored-by: sternenseemann <sternenseemann@systemli.org>