apache-airflow: 2.4.1 -> 2.4.3
Includes a patch to revert an upstream commit in order to get yarn to fetch
packages without errors.

Also, use the more correct forceFetchGit=true rather than leaveDotGit=true to
fetch sources from github using the git protocol (needed to get tests, which
aren't present in the tarball). leaveDotGit is not fully deterministic and this
was causing sha256 mismatches for some people.

Fixes #201763