python3Packages.pybind11: fix cross-compilation with setup hook
pybind11's CMake module uses the Python sysconfig module to find the Python
site-packages and include directories. sysconfig returns the wrong values
(from build Python) when cross-compiling, so this patch adds a setup hook to
pass flags to CMake to override the sysconfig results.

Note that the setup hook references build Python, and therefore adds build
Python to pybind11's runtime closure. This is not a problem because pybind11
is a header only library and should never end up as part of another package's
runtime closure.