waagent: update to 2.8.0.11 (#206974)
The agent has not been updated for a very long time.  In addition to
updating to the newest tagged version the change creates a package for
it.

The existing version has issues with the new python2.7 package not
containing crypt.so file. And the commit
6910a4eea0038728a2f10ce84122806f2cb6b170 I believe introduced
regression that caused the shebang to not be updated.