nixos/lib/make-disk-image.nix: fix to reserve correct number of inodes
If disk size computation is set to auto in nixos/lib/make-disk-image.nix, it computes the required space for the disk image by measuring the closure size that is to be copied onto the image. This computation takes into account the number of required inodes. However, when creating the disk image, the number of required inodes is not explicitly specified, so that mkfs.${fsType} uses the default inode ratio (which seems to be 1 inode for every 16384 bytes for ext4). This might be too low; see for example the failing example flake given in #292737. In that case, builds may fail with the error cptofs failed. diskSize might be too small for closure. during the call to cp2fs.

To avoid this, this pull request sets the number of inodes (multiplied with a factor compute_fudge that was already used in the storage calculation to add some margin) explicitly during the call to mkfs.${fsType} using the -N flag of mkfs.ext4. It also adds some additional logging output concerning the number of calculated inodes.

Co-authored-by: chessai <chessai@arista.com>