lix: only use LTO with GCC
Enabling LTO causes test failures when compiling for x86_64-darwin (at
least under Rosetta 2). The Nix package only enables LTO when using GCC,
so we'll copy that for now. I don't feel safe leaving it on for
aarch64-darwin unless we can understand why it's failing on
x86_64-darwin in case this is an indication of UB.

Fixes #337036.