emacs: disable native compilation on Darwin
The duplicate `rpath` issue prevents running Emacs on macOS 15.4 and
also breaks the build on that platform, blocking the channels. From
asking around, it seems like native compilation shouldn’t have
such a huge impact on performance as to not make this worth it as
a temporary hack to unblock security fixes and give us more time to
sort out a proper solution.