(partial: 0d079338b2f5) nixos/tests/installer: introduce `lix-simple` variant
An installer tests exacerbates the distribution packaging but in the
case of NixOS: the Nix package manager implementation.

As part of our classical release management process, the Lix team tests
whether a NixOS system installs just fine with Lix or not.

To avoid bloating the CI needlessly and keeping it simple, we only
introduce it on the simple variant and give a general way to pipe a Nix
implementation inside of an installer test.

Change-Id: I781da14475867dc2d946b740bad10af5de79ec5a
Signed-off-by: Raito Bezarius <masterancpp@gmail.com>